classifier/requirements.txt
python-telegram-bot>=20.7
requests>=2.31.0
Pillow>=10.0.0

torch>=2.1.0
numpy>=1.23.0
opencv-python>=4.8.0.76
kornia>=0.7.0


